stats-and-figures/requirements.txt
matplotlib
pandas
psycopg[binary]
seaborn
python-dotenv
scikit-learn
IPython